Fixed templates names and merged jira unit tests
