+++ app/tests/api/api_v1/test_jira.py
@@ -26,3 +26,26 @@ def test_jira_comment_default(
                     headers=api_key_headers)
 
     assert r.status_code == 200
+
+
+@mock.patch('jira.JIRA.create_issue')
+def test_jira_issue_default(
+        mock_post, client: TestClient, api_key_headers: Dict[str, str]) -> None:
+    params = {
+        "project_key": "JRA",
+        "issue_type": "Bug",
+        "issue_summary": "Test"
+    }
+    _body = {
+        "body": "SAMPLE MESSAGE"
+    }
+    api_key_headers.update({"url": "https://jira.atlassian.com"})
+
+
+    r = client.post(f"{settings.API_V1_STR}/jira/new_issue",
+                    json=_body,
+                    params=params,
+                    allow_redirects=True,
+                    headers=api_key_headers)
+
+    assert r.status_code == 200
