+++ docs/sample/chat_default.json
@@ -1,3 +1,3 @@
 {
-	  "body": "SAMPLE MESSAGE\nSAMPLE DETAILS"
+	"body": "SAMPLE MESSAGE\nSAMPLE DETAILS"
 }
