+++ docs/sample/jira_default.json
@@ -1,0 +1,3 @@
+{
+	"body": "This is an example - notify-service"
+}
