+++ docs/sample/jira_list.json
@@ -1,7 +1,7 @@
 {
 	"body":
 	{
-		"body": "This is an example - header",
-        "items": ["Item 1", "Item 2", "Item 3", "Item 4"]
+		"header": "This is an example - header",
+		"items": ["Item 1", "Item 2", "Item 3", "Item 4"]
 	}
 }
